% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_covariate_output.R
\name{construct.covariate.output}
\alias{construct.covariate.output}
\title{Append requested covariate information to a partially
constructed model matrix}
\usage{
construct.covariate.output(
  output.df,
  phenotype.data,
  phenotype.config,
  eigenvectors,
  analysis.config,
  analysis.name,
  collapse.limit
)
}
\arguments{
\item{output.df}{data frame; partially constructed model
matrix with, minimally, minimally, a single first column
containing subject IDs}

\item{phenotype.data}{data frame; loaded phenotype
data with informative headers. needs minimally a subject ID
column (determined based on configuration) and columns
corresponding to requested covariates}

\item{phenotype.config}{list; loaded phenotype
configuration yaml. needs minimally a variable flagged
as subject ID (see documentation of process.phenotypes)}

\item{eigenvectors}{data frame; loaded eigenvector data.
first column is subject ID; next N-2 columns are eigenvectors
starting with PC1; last column is a categorical variable
flagging populations for use with PC projection. by convention,
last column entries should be either "unrel" or "rel"
(though the "rel" category is allowed to be absent)}

\item{analysis.config}{list; loaded analysis-specific
configuration data. used to collect requested covariate
names and number of principal components. see documentation
of gwas-analysis-pipeline for more information}

\item{analysis.name}{character vector; unique name of analysis
currently being processed. this should correspond to one of
the block names under the "analyses:" tag in the analysis
configuration}

\item{collapse.limit}{integer; number of observations of a
category level above which the category is allowed to be a
distinct binary predictor in binarized categorical covariates.
many downstream models will fail if categories with extremely
low observation counts (e.g. 10 or fewer, or something like that)
are included in the regression model. levels for any categorical
predictor will be collapsed into a single combined predictor
that serves as a catchall for low count categories. note that
there is no guarantee that the combined variable will have
more observations than this limit, so it still may not converge.
in a future patch, this behavior may be modified to set subjects
falling into this catchall to NA}
}
\value{
input data frame with requested covariate columns appended
}
\description{
given a partially constructed output data frame
with selected subject IDs in the first column, use analysis
and phenotype configuration information to construct a
model matrix
}
\details{
this should aggregate information into a model matrix
regardless of whatever else may be present in the dataset
originally, so it can be used to build on top of an existing
phenotype data frame if desired. eigenvectors are provided
as input but only appended if a valid number of principal
components have been requested in analysis configuration.
categorical covariates are automatically converted into
binary dummies with category collapse as appropriate
}
