% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_for_plink.R
\name{format.for.plink}
\alias{format.for.plink}
\title{Apply formatting to data frame output in preparation
for use with plink2}
\usage{
\method{format}{for.plink}(output.df)
}
\arguments{
\item{output.df}{data frame; partially formatted model matrix}
}
\value{
reformatted input with plink-specific modifications
}
\description{
given a prepared model matrix, this function
will add a leading (all "0") FID column, and drop subjects
with duplicated IIDs, likely due to strange discrepancies
with ID mapping
}
\details{
this function is basically a stub, as the upstream
processing is largely performed with the idea that downstream
tools will accept plink-style format. other formatters for
other software tools (e.g. snptest) may be more extensive
}
