% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_trait_files.R
\name{get.lvl.names.labels}
\alias{get.lvl.names.labels}
\title{Get categorical level mapping keys and primary names from configuration data.}
\usage{
get.lvl.names.labels(phenotype.config, phenotype.shared.models, varname)
}
\arguments{
\item{phenotype.config}{Character vector name of phenotype configuration file.}

\item{phenotype.shared.models}{Character vector name of shared models configuration file.}

\item{varname}{Character vector standardized name of variable of interest.}
}
\value{
List, first entry "names" the mapping keys from the corresponding "levels" list,
and second entry "labels" the primary alias for each categorical level.
}
\description{
Get categorical level mapping keys and primary names from configuration data.
}
