% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pcs.R
\name{get.pcs}
\alias{get.pcs}
\title{Extract PCs from a smartpca-format eigenvector file}
\usage{
get.pcs(df, n.pcs, ids)
}
\arguments{
\item{df}{data frame; input eigenvector data in smartpca
style format, subject IDs in first column and set as rownames}

\item{n.pcs}{integer; number of requested principal components
for extraction}

\item{ids}{character vector; subject IDs for which PC data
are requested, in desired output order}
}
\value{
requested PC data, in correct order, as a data frame
}
\description{
this simple function just inspects an
eigenvector data frame for valid sets of entries and
emits the results for binding into a model matrix
}
\details{
this may not really justify its own function,
but the mapping is just finicky enough to suggest
it may benefit from specific testing. note that subjects
in query may be absent from PC data, and will simply
return NA for all PCs
}
