% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_and_process_eigenvectors.R
\name{load.and.process.eigenvectors}
\alias{load.and.process.eigenvectors}
\title{Read smartpca-format eigenvectors from file and apply
transformations for downstream use}
\usage{
load.and.process.eigenvectors(eigenvectors)
}
\arguments{
\item{eigenvectors}{character vector; name of input
smartpca eigenvector file}
}
\value{
loaded and transformed smartpca eigenvector data
}
\description{
this function expects raw output from smartpca,
and applies sufficient transformations to prevent plink2
from overly complaining about their distribution
}
\details{
the transformation requested by plink2 is quite
standard for regression models, so for the moment it is
unconditionally enabled. however, it is quite possible some
people may eventually object to this happening for some tools,
so future patches may introduce the ability to disable
the transformation on user request
}
