% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_trait_files.R
\name{map.factor.levels}
\alias{map.factor.levels}
\title{Before merging files, resolve level mappings in factor variables to match
the mappings in the original file.}
\usage{
map.factor.levels(df, phenotype.config, phenotype.shared.models, index)
}
\arguments{
\item{df}{Data frame of constructed model matrix.}

\item{phenotype.config}{Character vector of input phenotype configuration files.}

\item{phenotype.shared.models}{Character vector of input shared models configuration files.}

\item{index}{Integer or numeric index of currently merging file.}
}
\value{
Data frame of input data with factor variables remapped as needed.
}
\description{
Before merging files, resolve level mappings in factor variables to match
the mappings in the original file.
}
