% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_trait_files.R
\name{map.lvl.data}
\alias{map.lvl.data}
\title{Allow merging of categoricals with identical primary aliases but different
arbitrary mapping keys.}
\usage{
map.lvl.data(vec, ref.data, alt.data)
}
\arguments{
\item{vec}{Factor of input data that requires mapping.}

\item{ref.data}{List of names and labels for reference categorical configuration.}

\item{alt.data}{List of names and labels for merging categorical configuration.}
}
\value{
Factor of input data remapped as needed.
}
\description{
Allow merging of categoricals with identical primary aliases but different
arbitrary mapping keys.
}
