% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_controls.R
\name{match.controls}
\alias{match.controls}
\title{Perform ad hoc control matching for a phenotype.}
\usage{
match.controls(
  phenotype.file,
  covariate.file,
  output.file,
  n.pcs.match,
  n.controls.per.case
)
}
\arguments{
\item{phenotype.file}{Character name of input phenotype dataset.}

\item{covariate.file}{Character name of input covariate dataset.}

\item{output.file}{Character name of output phenotype dataset.}

\item{n.pcs.match}{Integer or numeric number of principal components
to use in matching model.}

\item{n.controls.per.case}{Integer or numeric number of controls to
include in the output dataset per case included.}
}
\description{
Given a pregenerated set of cases and controls, and their associated
principal components, perform control matching for cases. Attempts
to reach a fixed ratio of controls per included case.
}
