% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_variable.R
\name{transform.variable}
\alias{transform.variable}
\title{Apply requested transformation to a variable}
\usage{
\method{transform}{variable}(vec, transform.type, transform.strat)
}
\arguments{
\item{vec}{numeric vector; input data for transformation}

\item{transform.type}{character vector; name of requested transformation}

\item{transform.strat}{list; set of variables on which to stratify
transformation, when relevant}
}
\value{
transformed input vector
}
\description{
Apply requested transformation to a variable
}
